*: Update version to 0.0.1
The correct version should be reflected in the Cargo.toml file. When a
release is published, the crate will automatically have the same version
as the version in the Cargo.toml file. Since we are planning to release
an alpha 0.0.1 version, this should be the version in the Cargo.toml
file.

Signed-off-by: Dan Sover <dan.sover@avalabs.org>