feat: Return database creation errors through FFI (#945)
This returns errors from database creation across the FFI layer rather
than panicking in `expect` clauses.

I also wanted to test the case that prompted this (metric port errors),
but I couldn't cause it to fail in CI by opening a port with
`net.Listen`. If there's a way to consume the port, that would be really
helpful