chore: Set up for publishing to crates.io (#962)
It's been a while, but I'd like to get crates.io published. These are
all the changes required to do that.

Also, the authors have been updated to correctly reflect the authors for
the latest release. We'll still need to publish a new revision to get
this on crates.io.