build: move lints to the workspace (#957)
Previously, the lints were defined in each crate's `Cargo.toml`. This
commit moves the lints to the workspace's `Cargo.toml` file. This allows
for a more consistent linting configuration across all crates in the
workspace and simplifies the management of lint rules. This change also
ensures that all crates in the workspace adhere to the same linting
standards, improving code quality and maintainability.

New crates added to the workspace via `cargo new` will automatically
inherit the linting rules defined in the workspace's `Cargo.toml`.
However, manually crafted crates will need to ensure the stanza

```toml
[lints]
workspace = true
```

is included in `Cargo.toml`to benefit from the workspace-wide linting
rules.

Caveats:
- This change adds an excessive amount of warnings because
`clippy::pedantic` is very easy to violate and is now enabled workspace
wide. This is also why I included pedantic lints with a lower priority,
so that any individual lint within pedantic can be separately disabled
with an `allow`.
- Packages cannot override the workspace lints within their
`Cargo.toml`. This can be worked around by adding the lint rule change
via a crate attribute in the crate's `lib.rs` or `main.rs`. See the
`ffi` crate for an example of this workaround.

Resolves #951