chore: Simplify + cleanup generate_cgo script (#979)
- Simplify CGO block processing with state machine approach
- Correct isCGODirective to ensure each line has an actual directive
- Remove duplicate CGO LDFLAGS from memory.go, proposal.go, revision.go
- Avoid writing the file if it is unchanged
- Cleaner error reporting
- Improved documentation

The build probably still generated some warnings because the release
directory didn't exist, and yet it was being added to LDFLAGS in a few
other go files. This had the side effect of making the math library get
included on linux. We should always just include the math library for
dev builds.

---------

Co-authored-by: Aaron Buchwald <aaron.buchwald56@gmail.com>