refactor(deps): move duplicates to workspace (#1002)
This PR closes #929 by utilizing the `workspace.dependencies` table to
remove duplicate dependencies across crates.

## How

Across the member crates of the workspace, I identified dependencies
which appeared more than once and specified it in the workspace
dependencies table.

In the case of two instances of a dependency with different minor
versions, I default to using the greater minor version in the workspace.