chore: share workspace metadata and packages (#1020)
Many of the package fields can be declared at the workspace level and
then reused by the individual packages. This reduces duplication and
drift.

Additionally, moving the workspace packages to be listed as workspace
dependencies instead of strewn throughout the multiple Cargo.toml files
makes it easier to bump version numbers across the workspace as they are
all in one location. If we ever want to go back to individually set
version, we can swap back the `version = "..."` line in each package as
needed.

## Testing

It was easy to check what cargo would complain about before publishing
by using `cargo publish --dry-run -p <pkg>`. This was helpful to verify
that the changes to the `license-file` and `readme` fields would be
correctly transformed.

e.g.,

```console
$ cargo publish --dry-run -p firewood-storage
warning: crate firewood-storage@0.0.7 already exists on crates.io index
...
warning: aborting upload due to dry run
```