fix: remove dependency on serde (#1066)
The way we used serde when serializing free areas but not other nodes
was inconsistent and caused an issue were we were expecting data to be
encoded one way but in actuality it was encoded another way.

This commit removes the dependency on serde for serialization in the
storage crate, specifically in the `FreeArea` serialization. This
ensures that all serialization is done manually, which is consistent and
avoids the issues we were facing with serde's encoding.

Fixes: #831 #1057