feat!: auto open or create with truncate (#1064)
It's cumbersome (and more prone to TOCTOU issues) for users to validate
whether a file exists, and if it does, set a flag. This should be
handled internally. The `truncate` config now more closely resembles
it's actual meaning. If the file already exists, it will be wiped and
re-created.

There are unit tests in Rust and in Go for this.

Close #1039