fix: root_hash_reversed_deletions duplicate keys (#1076)
This test inserts a bunch of keys and then verifies that the resulting
hashes are the same if we delete them in the reverse order. However, the
test is generating a duplicate key, which will not reproduce the
original hash set.

Changed the test to remove any duplicate keys found. It's okay if the
data is the same.

Now passes with the errant key from #1075

Also fixed the merkle dump generator, which I used to debug this. A
bracket was missing so graphviz was failing to parse the file. Also
shortened the hashes as they made the nodes look much bigger, and added
a new line. See attached merkle dump for the new format.