chore: Update CODEOWNERS (#1080)
I noticed I was not auto-added to reviews for the ffi and .github
directories, and I discovered that this is because the wildcard rule was
being ignored if a more specific rule matched.

I have updated the `CODEOWNERS` file to ensure that Ron and I are added
to all reviews.