feat: Update RangeProof structure (#1136)
While implementing the range proof verification, I realized that having
an Option over a byte range is not ideal to work with. This meant that I
needed to do `proof.is_none_or(|p| p.is_empty())` every time I wanted to
use say a proof was empty. It was easier to use `proof.is_empty()`
everywhere. The inverse was also true where when I wanted to use a
proof, I first needed to destructure it and then check if it was empty.

```rust
let proof = match proof { Some(proof) if !proof.is_empty() => Some(proof), None };
```

Was in many places in my code.

Also updated is the method signature and documentation for
`verify_range_proof` to match how it will be used in the future.