ci: run CI on macOS (#1168)
As noted in #1129 and #1135, there are some inconsistencies with builds
for macOS and Linux. This commit adds a macOS runner to the CI workflow
to ensure that the lints and tests pass on macOS as well.

Instead of splitting the matrix to hoist the os separately, I simply
added a new element to the include block. This way the GitHub blocks on
merging Pull Requests won't be confused by the change in matrix values
(it matches on all of them literally).

The cache key does not need a token for `macos` because the full cache
key already includes the kernel type as a differentiator.

E.g.,

```log
Cache hit for: v0-rust-debug-no-features-Linux-x64-6dc3ef10-b549d219
```

versus

```log
Cache hit for: v0-rust-debug-no-features-macos-Darwin-arm64-c5c5f0bd-b549d219
```

The extra `macos` suffix is redundant and unnecessary.

Fixes #1129