feat(ffi): add BorrowedBytes type (#1174)
Add a `BorrowedSlice<_>` abstraction to provide a more ergonomic method
of providing sliced data to Rust from Go.

Go and Rust are similar in that they use ptr/len slices for all slice
types, not just byte slices. Additionaly, Go and Rust represent strings
as byte slices, meaning the conversion to and from CStrings is
unnecessary.

A future change will include a `OwnedSlice<_>` type that Rust will
provide back to Go callers for the same purpose, allowing Go to take
ownership of the data that Rust returns. This will complete an
abstraction over borrowed and owned slices that can be used for future
FFI work.