feat(ffi-refactor): Introduce VoidResult and panic handlers (3/8) (#1224)
As part of the larger refactor, this adds a `VoidResult` type to the FFI
library that will
be returned on methods that previously returned `void` (or unit).
Additionally added are two traits used by `invoke` and
`invoke_with_handle` to automatically handle the common null pointer
error and capture any panics from rust. These traits are extensively
used in upcoming pull requests.

Fixes: #1128

Depends On: #1223