feat(proofs)!: disable `ValueDigest::Hash` for ethhash (#1269)
As mentioned in #1267, the `ValueDigest::Hash` variant is not compatible
with the ethereum hash function. This disables it when the `ethhash`
feature is enabled to clear up some confusion until a more in-depth
refactor is done.

There is a potential advantage to simply leaving it this way. With the
conditional inclusion of the second variant, `ValueDigest` becomes a
thin new-type wrapper when `ethhash` is enabled because it contains only
one variant. I.e.,

```rust
let ValueDigest::Value(value) = value_digest;
```

is correct when `ethhash` is enabled because the let binding is
irrefutable.