fix: use `count` instead of `size_hint` (#1268)
`size_hint` is not guaranteed to be accurate and in some cases where we
use it, it may not be. This can potentially lead to silent bugs due to
incorrect assumptions about the number of elements in an iterator.

The `count` method on iterators _may_ be suboptimal and try enumerating
each item in the iterator; however, most iterator implementations are
smarter than that and will shortcut evaluating `count` and return a more
optimal response. E.g.,
[`core::slice::Iter`](https://doc.rust-lang.org/stable/src/core/slice/iter.rs.html#1376-1379)
or
[`alloc::vec::IntoIter`](https://doc.rust-lang.org/stable/src/alloc/vec/into_iter.rs.html#256-259)
and most iterator adaptors take advantage of this as well such as
[`core::iter::Chain`](https://doc.rust-lang.org/stable/src/core/iter/adapters/chain.rs.html#89-99)
and
[`core::iter::Copied`](https://doc.rust-lang.org/stable/src/core/iter/adapters/copied.rs.html#86-88).

I also refactored nibbles_to_eth_compact to be clearer and take
advantage of newer rust features to remove some redundant code.